+++ lib/extensio.dart
@@ -1,2 +1,6 @@
 export 'src/string_extensio.dart';
 export 'src/int_extensio.dart';
+export 'src/text_extensio.dart';
+export 'src/border_radius_extensio.dart';
+export 'src/build_context_extensio.dart';
+export 'src/bool_extensio.dart';