+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.3
+
+* Bool extensions added.
+* BorderRadius extensions added.
+* BuildContext extensions added.
+* TextStyle extensions added.
+
 ## 0.0.2
 
 * Integer extensions added.